// frontend/src/models/createCustomerDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * 企账通SCRM API
 * 企账通SCRM系统API文档
 * OpenAPI spec version: 1.0
 */

export interface CreateCustomerDto {
  /** 客户名称 */
  name: string;
  /** 联系人姓名 */
  contactName: string;
  /** 联系电话 */
  contactPhone: string;
  /** 联系邮箱 */
  contactEmail?: string;
  /** 公司名称 */
  companyName?: string;
  /** 地址 */
  address?: string;
  /** 客户等级: 0:潜在 1:意向 2:正式 3:VIP */
  customerLevel: number;
  /** 来源渠道 */
  sourceChannel?: number;
  /** 跟进人ID */
  followUserId?: string;
  /** 标签(JSON数组) */
  tags?: string;
  /** 备注 */
  remark?: string;
}